requirements.txt
absl-py==1.0.0
aiohttp==3.7.4.post0
appnope==0.1.2
APScheduler==3.6.3
asttokens==2.0.5
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==21.4.0
Automat==20.2.0
autopep8==1.6.0
backcall==0.2.0
beautifulsoup4==4.9.1
black==21.12b0
cachetools==4.2.2
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
configparser==5.2.0
constantly==15.1.0
crayons==0.4.0
cryptography==36.0.1
cssselect==1.1.0
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
entrypoints==0.3
et-xmlfile==1.1.0
executing==0.8.2
flatbuffers==2.0
fonttools==4.28.5
gast==0.5.3
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.43.0
h11==0.13.0
h2==3.2.0
h5py==3.6.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
ipykernel==6.7.0
ipython==8.0.1
ipython-genutils==0.2.0
ipython-sql==0.4.0
itemadapter==0.4.0
itemloaders==1.0.4
jedi==0.18.1
jmespath==0.10.0
joblib==1.1.0
jupyter-client==7.1.2
jupyter-core==4.9.1
keras==2.8.0rc1
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
lxml==4.7.1
Markdown==3.3.6
matplotlib==3.5.1
matplotlib-inline==0.1.3
multidict==6.0.2
mypy-extensions==0.4.3
nest-asyncio==1.5.4
networkx==2.6.3
numpy==1.22.1
oauthlib==3.1.1
openpyxl==3.0.9
opt-einsum==3.3.0
outcome==1.1.0
packaging==21.3
pandas==1.4.0
parsel==1.6.0
parso==0.8.3
pathspec==0.9.0
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.0
platformdirs==2.4.1
plotly==5.5.0
plotly-geo==1.0.0
prettytable==0.7.2
priority==1.3.0
prompt-toolkit==3.0.24
Protego==0.1.16
protobuf==3.19.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
PyDispatcher==2.0.5
Pygments==2.11.2
pyOpenSSL==21.0.0
pyparsing==3.0.7
pyTelegramBotAPI==4.4.0
python-dateutil==2.8.2
python-sec==0.1.6
python-telegram-bot==13.7
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pyzmq==22.3.0
queuelib==1.6.2
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.8
scikit-learn==1.0.2
scipy==1.7.3
Scrapy==2.5.1
seaborn==0.11.2
sec-api==1.0.8
secedgar==0.3.3
selenium==4.1.0
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
SQLAlchemy==1.4.31
sqlparse==0.4.2
stack-data==0.1.4
statsmodels==0.13.1
telebot==0.0.4
tenacity==8.0.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0rc0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.0.0
toml==0.10.2
tomli==1.2.3
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
trio==0.19.0
trio-websocket==0.9.2
Twisted==21.7.0
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.5
w3lib==1.22.0
wcwidth==0.2.5
webdriver-manager==3.5.2
Werkzeug==2.0.2
wrapt==1.13.3
wsproto==1.0.0
yarl==1.7.2
zope.interface==5.4.0
